requirements.txt
burnysc2>=7.1.0
